Refactor context to live in its own folder and separate hooks
